Use Python multiprocessing module to speed up file-scan step
Even though this program is IO-bound, HDF5 is not designed to support
multithreading efficiently.  Therefore, using threads will not improve
our performance.  But, if we have operations that can be run in
subprocesses, this may be viable.

This version of preshape.py includes multi-process execution of input
file scanning, which should dramatically improve the performance of this
step.  We can experiment with it on Perlmutter to see if it helps.