Example config files for generating SLURM experiments
These are some prototype config files for generating Perturbo
configuration for SLURM on NERSC Perlmutter.  We don't have a program
to use these yet, but they illustrate the idea of what I want to build.

The .tmpl files are templates from which config files can be generated.
Since Perturbo needs to run through three steps to set up for a
simulation, the mp_conf.json file specifies the three sets of files to
generate from the templates, and what their values should be.  It also
has a command for running each step.

The tool will be able to generate the specified files, and optionally
kick off the SLURM jobs as well.