Dry-run support; better control of multiprocessing implementation
The previous version of preshape was not smart about capping the number
of subprocesses running at any given time.  For large sets of files,
preshape would completely saturate the IO subsystem and make the
computer unresponsive.

This version resolves that issue by enforcing a maximum number of
processes that may run at any given time.  Reshape operations take a bit
more time to complete, but the system remains usable.  The default
maximum is 50 subprocesses, but this can now be controlled with the -M
(or --max-processes) argument.

A -n (or --dryrun) option was also added to just run the input-scanning
portion of the program without generating any output files.

A -q (or --quiet) option was also added to turn off unnecessary output,
so that preshape can be put into the background without being quite so
noisy in the terminal.