Add --set and --foreach command-line arguments
These arguments make the `pertool generate` command much more powerful,
as it can now take a list of values for a given argument and generate a
set of output directories, one for each value in the list.  It can
support multiple `--foreach` arguments, and will generate all possible
combinations of the listed values.

I think the implementation is not quite complete, since we still need
the ability to copy files in directories etc., but it is pretty close
to being usable for the Perturbo performance tests.