Fix copy-file/symlink bugs in pertool generate command
Using pertool to generate GaAs simulation configurations revealed a few
issues with how file-copying and symlinking was working.  The most
obvious issue was that symlinking was relative to the directory where
pertool was run from, as opposed to the target-directory being
generated, which meant the symlinks didn't point to the correct target.
This is now fixed.

Another issue that emerged is how to deal with subdirectories, where
some may have very large files (and therefore should not be copied), but
others may have small files (and it's therefore fine to copy them).
Right now pertool just symlinks all of them, since the primary
subdirectory will be the "/tmp" directory containing the eph-g2 files
generated for dynamics calculation, and we definitely don't want to copy
those files.